Pretty large initial commit.
Includes Server and Client made by Ryan, modified slightly by me. Still
a lot of comments and test prints in that file.

	Also the schema.xml file included with that server is out of
	date and must be updated before this app is ready

And app.py file made by Long and I.

The app.py file has the GUI.

It is currently not connected to either the Client or the Server files.
But it does generate a functional XML document.